backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==4.2.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
brave-search==0.1.8
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
elevenlabs==1.54.0
entrypoints==0.4
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
Flask==2.2.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-experimental==0.3.4
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pytest==8.3.5
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytubefix==8.12.2
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
streamlit==1.43.1
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==2.3.8
yarl==1.18.3
zstandard==0.23.0